+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## [Unreleased]
 ### Added
 * `adjuster.number().strict()`
+* `adjuster.string().strict()`
 
 ## [0.14.0] - 2018-08-11
 ### Added